Refactor: reconstruction of AbstractFutureState and AggregateFutureState AtomicReferenceFieldUpdater and AtomicIntegerFieldUpdater way of creating and using
In AbstractFutureState, waiters, listeners and the value of AtomicReferenceFieldUpdater way create multiline comment from single-line comments). In the AggregateFutureState, the static initialization block was restructured, and the creation method of the SafeAtomicHelper was changed to passing in seenExceptions and the remaining updater. Meanwhile, the implementations of the getOrInitSeenExceptions and decrementRemainingAndGet methods have been optimized to ensure thread safety. In addition, removed the seenExceptionsUpdaterFromWithinAggregateFutureState and remainingCountUpdaterFromWithinAggregateFutureState method.